+++ src/Modules/DynamicContent/Policies/DynamicContentTypePolicy.php
@@ -18,26 +18,26 @@ class DynamicContentTypePolicy
 {
     public function viewAny( Authenticatable $user ): bool
     {
-        return $user->can( applyFilters( 'dynamicContent.viewAny', 'manage_dynamic_content' ) );
+        return $user->can( applyFilters( 'ap.cmsFramework.abilities.dynamicContent.viewAny', 'manage_dynamic_content' ) );
     }
 
     public function view( Authenticatable $user, DynamicContentType $type ): bool
     {
-        return $user->can( applyFilters( 'dynamicContent.view', 'manage_dynamic_content', $type ) );
+        return $user->can( applyFilters( 'ap.cmsFramework.abilities.dynamicContent.view', 'manage_dynamic_content', $type ) );
     }
 
     public function create( Authenticatable $user ): bool
     {
-        return $user->can( applyFilters( 'dynamicContent.create', 'manage_dynamic_content' ) );
+        return $user->can( applyFilters( 'ap.cmsFramework.abilities.dynamicContent.create', 'manage_dynamic_content' ) );
     }
 
     public function update( Authenticatable $user, DynamicContentType $type ): bool
     {
-        return $user->can( applyFilters( 'dynamicContent.update', 'manage_dynamic_content', $type ) );
+        return $user->can( applyFilters( 'ap.cmsFramework.abilities.dynamicContent.update', 'manage_dynamic_content', $type ) );
     }
 
     public function delete( Authenticatable $user, DynamicContentType $type ): bool
     {
-        return $user->can( applyFilters( 'dynamicContent.delete', 'manage_dynamic_content', $type ) );
+        return $user->can( applyFilters( 'ap.cmsFramework.abilities.dynamicContent.delete', 'manage_dynamic_content', $type ) );
     }
 }